moved events module to config
